[Refactor]Thumbnail, Pagination 파일 분리
